+++ docs/conf.py
@@ -24,7 +24,7 @@ copyright = '2020, HETDEX Data Release Team'
 author = 'Erin Mentuch Cooper'
 
 # The full version, including alpha/beta/rc tags
-release = '0.2'
+release = '0.5'
 
 
 # -- General configuration ---------------------------------------------------