+++ hetdex_api/detections.py
@@ -17,6 +17,11 @@ import os.path as op
 import numpy as np
 import tables as tb
 import copy
+
+np.warnings.filterwarnings("ignore")
+
+import matplotlib
+matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 
 from astropy.table import vstack, Table, Column, join
@@ -28,14 +33,23 @@ import speclite.filters
 
 from hetdex_api.survey import Survey
 from hetdex_api.config import HDRconfig
+from hetdex_api.mask import *
 
-np.warnings.filterwarnings("ignore")
 PYTHON_MAJOR_VERSION = sys.version_info[0]
 PYTHON_VERSION = sys.version_info
 
+try:
+    from hetdex_api.config import HDRconfig
+    LATEST_HDR_NAME = HDRconfig.LATEST_HDR_NAME
+    
+except Exception as e:
+    print("Warning! Cannot find or import HDRconfig from hetdex_api!!", e)
+    LATEST_HDR_NAME = "hdr2.1"
 
+    
 class Detections:
-    def __init__(self, survey="hdr2.1", catalog_type="lines", loadtable=True):
+    def __init__(self, survey=LATEST_HDR_NAME, catalog_type="lines",
+                 curated_version=None, loadtable=True):
         """
         Initialize the detection catalog class for a given data release
 
@@ -65,10 +79,18 @@ class Detections:
             print(catalog_type_options)
             return None
 
-        global config
-        config = HDRconfig(survey=survey)
+        # store to class
+        if curated_version is not None:
+            self.version = curated_version
+            self.loadtable = False
+            self.survey = 'hdr' + curated_version[0:3]
+        else:
+            self.version = None
+            self.survey = survey
+            self.loadtable = loadtable
 
-        self.survey = survey
+        global config
+        config = HDRconfig(survey=self.survey)
 
         if catalog_type == "lines":
             self.filename = config.detecth5
@@ -83,9 +105,35 @@ class Detections:
         self.hdfile = tb.open_file(self.filename, mode="r")
 
         # store to class
-        self.loadtable = loadtable
+        if curated_version is not None:
+            self.version = curated_version
+            self.loadtable = False
+            self.survey = 'hdr' + curated_version[0:3]
+        else:
+            self.survey = survey
+            self.loadtable = loadtable
 
-        if loadtable:
+        if self.version is not None:
+
+            try:
+                catfile = op.join( config.detect_dir,
+                                   'catalogs',
+                                   "detect_hdr" + self.version + ".fits")
+                det_table = Table.read( catfile )
+
+                for col in det_table.colnames:
+                    if isinstance( det_table[col][0], str ):
+                        setattr(self, col, np.array( det_table[col] ).astype(str) )
+                    else:
+                        setattr(self, col, np.array( det_table[col] ) )
+
+                self.vis_class = -1 * np.ones(np.size(self.detectid))
+                        
+            except:
+                print('Could not open curated catalog version: ' + self.version)
+                return None
+
+        elif self.loadtable:
             colnames = self.hdfile.root.Detections.colnames
             for name in colnames:
                 if isinstance(
@@ -103,6 +151,7 @@ class Detections:
 
             # add in the elixer probabilties and associated info:
             if self.survey == "hdr1" and catalog_type=='lines':
+
                 self.hdfile_elix = tb.open_file(config.elixerh5, mode="r")
                 colnames2 = self.hdfile_elix.root.Classifications.colnames
                 for name2 in colnames2:
@@ -160,10 +209,13 @@ class Detections:
                     print('No Elixer table found')
                     
             # also assign a field and some QA identifiers
-            self.field = np.chararray(np.size(self.detectid), 12)
+            self.field = np.chararray(np.size(self.detectid), 12, unicode=True)
             self.fwhm = np.zeros(np.size(self.detectid))
             if self.survey == "hdr1":
                 self.fluxlimit_4550 = np.zeros(np.size(self.detectid))
+            else:
+                self.fluxlimit_4540 = np.zeros(np.size(self.detectid))
+                
             self.throughput = np.zeros(np.size(self.detectid))
             self.n_ifu = np.zeros(np.size(self.detectid), dtype=int)
 
@@ -178,6 +230,10 @@ class Detections:
                     self.fluxlimit_4550[ix] = S.fluxlimit_4550[index] 
                 else:
                     self.fwhm[ix] = S.fwhm_virus[index]
+                try:
+                    self.fluxlimit_4540[ix] = S.fluxlimit_4540[index]
+                except:
+                    pass
                 self.throughput[ix] = S.response_4540[index]
                 self.n_ifu[ix] = S.n_ifu[index]
 
@@ -195,10 +251,10 @@ class Detections:
 
             self.vis_class = -1 * np.ones(np.size(self.detectid))
 
-            if survey == "hdr1":
+            if self.survey == "hdr1":
                 self.add_hetdex_gmag(loadpickle=True, picklefile=config.gmags)
             
-            if survey == "hdr1":
+            if self.survey == "hdr1":
                 if PYTHON_MAJOR_VERSION < 3:
                     self.plae_poii_hetdex_gmag = np.array(
                         pickle.load(open(config.plae_poii_hetdex_gmag, "rb"))
@@ -219,6 +275,7 @@ class Detections:
         # set the SkyCoords
         self.coords = SkyCoord(self.ra * u.degree, self.dec * u.degree, frame="icrs")
 
+            
     def __getitem__(self, indx):
         """ 
         This allows for slicing of the Detections class
@@ -268,8 +325,12 @@ class Detections:
         else:
             mask1 = self.remove_bad_amps()
             mask2 = self.remove_bad_detects()
-
-            mask = mask1 * mask2
+            mask3 = self.remove_bright_stuff(gmagcut)
+            mask4 = self.remove_bad_pix()
+            mask5 = self.remove_shots()
+            mask6 = self.remove_meteors()
+            
+            mask = mask1 * mask2 * mask3 * mask4 * mask5 * mask6
             
         return self[mask]
 
@@ -469,8 +530,6 @@ class Detections:
             maskdet = self.detectid == baddet
             mask = np.logical_or(mask, maskdet)
 
-        self.vis_class[mask] = -2
-
         return np.invert(mask)
 
     def remove_bad_amps(self):
@@ -518,12 +577,34 @@ class Detections:
 
             return np.logical_not(mask)
         else:
+
+            # first read in amp_flag.fits file
             badamps = Table.read(config.badamp)
+            
             det_table = self.return_astropy_table()
+            
             join_tab = join(det_table, badamps, keys=['shotid','multiframe'], join_type='left')
-            mask = join_tab['flag'] == 1
+            # this is needed to match with detection class object sorting
+            join_tab.sort('detectid')
+
+            mask1 = join_tab['flag'] != 0
 
             del det_table, join_tab
+            
+            # add in any newly found badamps that haven't made it into the
+            # amp_flag.fits file yet
+            
+            mask2 = np.zeros( np.size(self.detectid), dtype=bool)
+            
+            badamps2 = Table.read(config.badamp2, format='ascii')
+
+            for row in badamps2:
+                selmf = self.multiframe == row['multiframe']
+                seldate = (self.date >= row['date_start']) \
+                          * (self.date <= row['date_end'])
+                mask2 = np.logical_or(mask2, selmf*seldate)
+                
+            mask = mask1 * np.logical_not( mask2)
             
             return mask
 
@@ -551,24 +632,28 @@ class Detections:
         
         """
 
-        badpixlist = ascii.read(
-            config.badpix, names=["multiframe", "x1", "x2", "y1", "y2"]
-        )
-
-        mask = np.zeros(np.size(self.detectid), dtype=bool)
-
-        for row in badpixlist:
-            maskbadpix = (
-                (self.multiframe == row["multiframe"])
-                * (self.x_raw > row["x1"])
-                * (self.x_raw < row["x2"])
-                * (self.y_raw > row["y1"])
-                * (self.y_raw < row["y2"])
+        if True:
+            badpixlist = ascii.read(
+                config.badpix, names=["multiframe", "x1", "x2", "y1", "y2"]
             )
-            mask = np.logical_or(maskbadpix, mask)
 
-        self.vis_class[mask] = 0
+            mask = np.zeros(np.size(self.detectid), dtype=bool)
 
+            for row in badpixlist:
+                maskbadpix = (
+                    (self.multiframe == row["multiframe"])
+                    * (self.x_raw >= row["x1"])
+                    * (self.x_raw <= row["x2"])
+                    * (self.y_raw >= row["y1"])
+                    * (self.y_raw <= row["y2"])
+                )
+                mask = np.logical_or(maskbadpix, mask)
+                
+            self.vis_class[mask] = 0
+
+        else: #except:
+            mask = np.zeros(np.size(self.detectid), dtype=bool)
+            
         return np.invert(mask)
 
     def remove_shots(self):
@@ -635,6 +720,26 @@ class Detections:
 
         return np.invert(mask)
 
+    def remove_meteors(self):
+        """
+        Returns boolean mask with detections landing on meteor
+        streaks masked. Use np.invert(mask) to find meteors
+        """
+        
+        global config
+
+        met_tab = Table.read(config.meteor, format='ascii')
+
+        mask = np.ones_like(self.detectid, dtype=bool)
+        
+        for row in met_tab:
+            sel_shot = np.where(self.shotid == row['shotid'])[0]
+            for idx in sel_shot:
+                coord = self.coords[idx]
+                mask[idx] = meteor_flag_from_coords(coord,
+                                                    row['shotid'])
+        return mask
+        
     def get_spectrum(self, detectid_i):
         """
         Grabs the 1D spectrum used to measure fitted parameters.
@@ -759,6 +864,11 @@ class Detections:
                     table[name] = getattr(self, name)
             except:
                 print('Could not add elixer columns')
+            try:
+                table.add_column(Column(self.fluxlimit_4540), index=3, name="fluxlimit_4540")
+            except:
+                print('Could not add average flux limit')
+                
         return table
 
     def save_spectrum(self, detectid_i, outfile=None):