+++ hetdex_api/survey.py
@@ -17,13 +17,20 @@ import numpy
 import copy
 import astropy.units as u
 from astropy.coordinates import SkyCoord
-from astropy.table import Table
+from astropy.table import Table, vstack
 
+import healpy as hp
 from hetdex_api.config import HDRconfig
+
+try:
+    LATEST_HDR_NAME = HDRconfig.LATEST_HDR_NAME
+except Exception as e:
+    print("Warning! Cannot find or import HDRconfig from hetdex_api!!", e)
+    LATEST_HDR_NAME = "hdr2.1"
 
 
 class Survey:
-    def __init__(self, survey='hdr2.1'):
+    def __init__(self, survey=LATEST_HDR_NAME):
         """
         Initialize the Survey class for a given data release
 
@@ -65,23 +72,26 @@ class Survey:
         self.coords = SkyCoord(self.ra * u.degree, self.dec * u.degree, frame="icrs")
 
         # append flux limits
-        if survey == 'hdr2.1':
-            
-            flim = Table.read('/data/05350/ecooper/hdr2.1/survey/average_one_sigma.txt',
-                              format='ascii', names=['datevobs', 'fluxlimit_4550'])
+        if survey == "hdr2.1":
+
+            flim = Table.read(
+                config.flim_avg,
+                format="ascii",
+                names=["datevobs", "col2", "fluxlimit_4540"],
+            )
             fluxlimit = []
             for datevobs in self.datevobs:
-                sel = flim['datevobs'] == datevobs
+                sel = flim["datevobs"] == datevobs
                 if np.sum(sel) == 1:
-                    fluxlimit.extend( flim['fluxlimit_4550'][sel] )
+                    fluxlimit.extend(flim["fluxlimit_4540"][sel])
                 elif np.sum(sel) > 1:
-                    print('Found two fluxlimits for ', datevobs)
-                    fluxlimit.extend( flim['fluxlimit_4550'][sel][0])
+                    print("Found two fluxlimits for ", datevobs)
+                    fluxlimit.extend(flim["fluxlimit_4540"][sel][0])
                 else:
-                    fluxlimit.append( np.nan )
-                    
-            self.fluxlimit_4550 = np.array(fluxlimit)
-                              
+                    fluxlimit.append(np.nan)
+
+            self.fluxlimit_4540 = np.array(fluxlimit)
+
     def __getitem__(self, indx):
         """ 
         This allows for slicing of the survey class
@@ -149,7 +159,7 @@ class Survey:
 
         notvalid = self.shotid < 20170000
         mask = mask | notvalid
-        
+
         return np.invert(mask)
 
     def get_shotlist(self, coords, radius=None, width=None, height=None):
@@ -207,7 +217,6 @@ class Survey:
 
         return self.shotid[idx]
 
-
     def return_astropy_table(self, return_good=True):
         """
         Function to return an astropy table that is machine readable
@@ -229,24 +238,24 @@ class Survey:
 
         good_shots = self.remove_shots()
 
-        survey_table['shot_flag'] = good_shots
+        survey_table["shot_flag"] = good_shots
 
-        survey_table['mjd'] = self.mjd[:,0]
-        survey_table['exptime'] = np.mean(self.exptime, axis=1)
-        survey_table['fluxlimit_4550'] = self.fluxlimit_4550
-        
+        survey_table["mjd"] = self.mjd[:, 0]
+        survey_table["exptime"] = np.mean(self.exptime, axis=1)
+        survey_table["fluxlimit_4540"] = self.fluxlimit_4540
+
         for col in survey_table.colnames:
             try:
                 if np.shape(survey_table[col])[1] == 3:
                     survey_table.remove_column(col)
             except:
                 pass
-                    
+
         if return_good:
             return survey_table[good_shots]
         else:
             return survey_table
-            
+
     def close(self):
         """
         Close the HDF5 file when you are done using
@@ -260,8 +269,9 @@ class Survey:
 
         self.hdfile.close()
 
+
 class FiberIndex:
-    def __init__(self, survey='hdr2.1', loadall=False):
+    def __init__(self, survey=LATEST_HDR_NAME, loadall=False):
         """
         Initialize the Fiber class for a given data release
         
@@ -270,15 +280,15 @@ class FiberIndex:
         survey : string
             Data release you would like to load, i.e., 'hdr1','HDR2'
             This is case insensitive.
-        
+
         Returns
         -------
         FiberIndex class object
         """
         self.survey = survey
-        
-        if self.survey == 'hdr1':
-            print('Sorry there is no FiberIndex for hdr1')
+
+        if self.survey == "hdr1":
+            print("Sorry there is no FiberIndex for hdr1")
             return None
 
         global config
@@ -289,11 +299,11 @@ class FiberIndex:
 
         if loadall:
             colnames = self.hdfile.root.FiberIndex.colnames
-            
+
             for name in colnames:
-            
+
                 if isinstance(
-                        getattr(self.hdfile.root.FiberIndex.cols, name)[0], np.bytes_
+                    getattr(self.hdfile.root.FiberIndex.cols, name)[0], np.bytes_
                 ):
                     setattr(
                         self,
@@ -301,14 +311,15 @@ class FiberIndex:
                         getattr(self.hdfile.root.FiberIndex.cols, name)[:].astype(str),
                     )
                 else:
-                    setattr(self, name,
-                            getattr(self.hdfile.root.FiberIndex.cols, name)[:])
-                    
-            self.coords = SkyCoord(self.ra[:] * u.degree,
-                                   self.dec[:] * u.degree,
-                                   frame="icrs")
+                    setattr(
+                        self, name, getattr(self.hdfile.root.FiberIndex.cols, name)[:]
+                    )
 
-    def query_region(self, coords, radius=3.*u.arcsec, shotid=None, astropy=True):
+            self.coords = SkyCoord(
+                self.ra[:] * u.degree, self.dec[:] * u.degree, frame="icrs"
+            )
+
+    def query_region(self, coords, radius=3.0 * u.arcsec, shotid=None):
         """
         Function to retrieve the indexes of the FiberIndex table
         for a specific region
@@ -324,29 +335,96 @@ class FiberIndex:
             radius you want to search. An astropy quantity object
         shotid
             Specific shotid (dtype=int) you want
-        """
-        
-        ra_obj = coords.ra.deg
-        ra_sep = radius.to(u.degree).value + 3./3600.
-        
-        tab = self.hdfile.root.FiberIndex
-        seltab = tab.read_where( '(ra < ra_obj + ra_sep) & (ra > (ra_obj - ra_sep))')
 
-        fibcoords = SkyCoord(seltab['ra'] * u.degree, seltab['dec'] * u.degree, frame='icrs')
+        Returns
+        -------
+        An astropy table of Fiber infomation in queried aperture
+        """
+
+        Nside = 2 ** 15
+
+        ra_obj = coords.ra.deg
+        dec_obj = coords.dec.deg
+
+        ra_sep = radius.to(u.degree).value + 3.0 / 3600.0
+
+        vec = hp.ang2vec(ra_obj, dec_obj, lonlat=True)
+
+        pix_region = hp.query_disc(Nside, vec, (ra_sep * np.pi / 180))
+
+        seltab = Table()
+        for hpix in pix_region:
+            h_tab = self.get_fib_from_hp(hpix)
+            seltab = vstack([seltab, h_tab])
+
+        fibcoords = SkyCoord(
+            seltab["ra"] * u.degree, seltab["dec"] * u.degree, frame="icrs"
+        )
 
         if shotid:
-            idx = (coords.separation(fibcoords) < radius) * (seltab['shotid'] == shotid)
+            idx = (coords.separation(fibcoords) < radius) * (seltab["shotid"] == shotid)
         else:
             idx = coords.separation(fibcoords) < radius
 
-        if astropy:
-            return Table(seltab[idx])
-        else:
-            return seltab[idx]
+        return seltab[idx]
 
+        
     def get_fib_from_hp(self, hp, astropy=True):
 
         if astropy:
-            return Table(self.hdfile.root.FiberIndex.read_where('healpix == hp'))
+            return Table(self.hdfile.root.FiberIndex.read_where("healpix == hp"))
         else:
-            return self.hdfile.root.FiberIndex.read_where('healpix == hp')
+            return self.hdfile.root.FiberIndex.read_where("healpix == hp")
+
+            
+    def get_closest_fiberid(self, coords, shotid=None, maxdistance=8.*u.arcsec):
+        """
+        Function to retrieve the closest fiberid in a shot
+        
+        Parameters
+        ----------
+        self
+            the FiberIndex class for a specific survey
+        coords
+            center coordinate you want to search. This should
+            be an astropy SkyCoord object
+        shotid
+            Specific shotid (dtype=int) you want
+        maxdistance
+            The max distance you want to search for a nearby fiber.
+            Default is 8.*u.arcsec 
+        
+        Returns
+        -------
+        fiberid
+            unique fiber identifier for the shot
+        
+        """
+
+        #start searching at small radius to search more efficiently
+        search = 2.*u.arcsec
+        
+        fiber_table = self.query_region(coords, radius=search, shotid=shotid)
+
+        if np.size(fiber_table) > 0:
+            pass
+        else:
+            fiber_table = self.query_region(coords, radius=maxdistance, shotid=shotid)
+
+        if np.size(fiber_table) > 0:
+            fibcoords = SkyCoord(
+                fiber_table["ra"] * u.degree, fiber_table["dec"] * u.degree, frame="icrs"
+            )
+            
+            idx = np.argmin(coords.separation(fibcoords))
+            fiberid = fiber_table['fiber_id'][idx]
+        
+            return fiber_table['fiber_id'][idx]
+        else:
+            return None
+
+    def close(self):
+        """
+        Close the hdfile when done
+        """
+        self.hdfile.close()