+++ ReleaseNotes.rst
@@ -1,6 +1,30 @@
 ``hetdex_api`` release notes
 ==========================
 
+Release 0.6
+-----------
+
+- Still release for HDR2.1.1
+
+- using healpy to speed up query speed on FiberIndex
+
+- small changes in docs for cleaner organization
+
+- catalog curation: added 82 new bad shots, a bad fiber and some bad amps, caught 960 charge trap issues.
+
+Release 0.5
+-----------
+
+- Release for HDR2.1.1 
+
+- Detections class function refine() now operational for HDR2.1
+
+- Added curated_version option for Detections class to pull stable catalog
+
+- small changes in docs for cleaner organization
+
+- catalog curation: added 82 new bad shots, a bad fiber and some bad amps, caught 960 charge trap issues. 
+
 Development version @ trunk
 ---------------------------